Fix for #1154
Partially roll back to pre-#1111 behavior.
We just make sure that the STRING shape is chosen when Shape.ANY (the default) is set on the annotation, but some other annotation attribute was also set (pattern, locale or timezone).
This way of fixing the issue has the added benefit of respecting the user config regarding the default serialization of dates when @JsonFormat(shape = Shape.ANY) is set on a property.