WIP: Allow more control over default creator property values
There are no javadocs or tests yet; this is to set up a pull request for
discussion.

This should be a backwards-compatible change.  A new method is being
added to ValueInstantiator, but by default it will call an old method
and have exactly the same old behavior.  The behavior only changes if and
when the application decides to override this new method in their own
ValueInstantiator.  (I suppose that if someone out there already made
their own ValueInstantiator subclass with this exact signature, this
change would not be backwards-compatible for them.  But that seems like
a very unlikely scenario.)

The hasParameter and getParameter methods of PropertyValueBuffer are not
used by jackson-databind itself.  Those are the methods that we expect
applications to use if they want finer-grained control over default
values.  Instead of eagerly fetching every default value (and possibly
throwing an exception for missing-but-required parameters), applications
can use hasParameter to check if each property is present, and then use
getParameter to fetch the known-present values one by one.

One weird thing about this change is that it exposes a class in the
deser.impl package (PropertyValueBuffer) in the API of the root deser
package (ValueInstantiator).  That seems "wrong" given the names of the
packages, but maybe no one cares?  We could make PropertyValueBuffer
implement an interface that exposes hasParameter, getParameter, and
getParameters, then use that interface in ValueInstantiator instead...
I suspect that no one would ever implement that interface except
PropertyValueBuffer, making the interface useless.

The end goal of this is to let Kotlin applications use that language's
support for default parameter values.  The same mechanism might work for
other JVM-based languages with default values, but I have only proved it
out with a Kotlin application.

Related discussion:
https://github.com/FasterXML/jackson-module-kotlin/issues/26