fix ObjectMapper: Bind types to TypeReference<> parameter
Previously functions making use of types passed via `TypeReference<T>`
did not bind return values explicitly to the referenced type.

This changes the signatures of those functions to make use of the
referenced type.

This will cause compile-time errors in code that worked before, such as:

    Long l = mapper.readValue(someString, new TypeReference<String>(){});

Which should fail as `String` cannot safely be cast to `Long`.