fix test: Fix several type errors in tests
The previous commit that properly bound return types of readValue when
using TypeReference instances to pass types highlighted several errors
in test cases.

Some of these I believe are genuine errors that worked due to a series
of issues (e.g. `Map::get()` is not type-safe due to legacy
compatiblity).

Also removes an assertTrue that is now guaranteed by the type
system (IntelliJ suggested optimising it to `assertTrue(true)` :))