Support READ_UNKNOWN_ENUM_VALUES_AS_NULL with @JsonCreator
The deserialization feature READ_UNKNOWN_ENUM_VALUES_AS_NULL allows enum
values that are not recognised to take a null value when parsed. Before
this commit, this deserialization feature did not work whenever an enum
had a method marked @JsonCreator, and the enum creator method itself had
to choose null in the case of unknown input.

With this change, if an enum creator method throws an
IllegalArgumentException then this is considered to be an unknown value
and (if READ_UNKNOWN_ENUM_VALUES_AS_NULL is active) then null will be
used.