Fix failing test for ObjectWriter when running from Canada
The test was using Locale.CANADA to create a new ObjectWriter and
then compare that it is a new one, and not the same ObjectWriter used
to create it. But when the default Locale is Locale.CANADA, then the
same ObjectWriter is returned and the fails.

This is then using a locale that is not the same as Locale.getDefault to
ensure the test is gonna pass no matter which locale is the default.