Fix failing tests when running from Canada (#2808)
* Fix failing test for ObjectWriter when running from Canada

The test was using Locale.CANADA to create a new ObjectWriter and
then compare that it is a new one, and not the same ObjectWriter used
to create it. But when the default Locale is Locale.CANADA, then the
same ObjectWriter is returned and the fails.

This is then using a locale that is not the same as Locale.getDefault to
ensure the test is gonna pass no matter which locale is the default.

* Enable PackageVersion.java generation before compiling code

This makes the code generation friendly to IDEs (such as IDEA) when
importing the project, and running the tests inside of it.