Fix TokenBuffer._copyBufferValue floating point values
regressed handling inside polymorphic types. The TokenBuffer could
hold values as BigDecimal objects which are heavy on heap, but
critically also cannot be used to deserialize to a less precise
target, e.g. double. This is the inverse problem of #2644 where
not enough precision was retained, and reintroduces precision loss
while buffering floating point contents.