Alternate approach for wildcards in TypeBindings
@pjfanning ok i've looked at #4149 and it'll be awkward. At least for `TypeFactory_mapType`, the fix would be easy, just add this case to fix the test failure:

```java
            case 1:
                TypeVariable<? extends Class<?>>[] variables = rawClass.getTypeParameters();
                if (variables.length == 2) {
                    String boundName = bindings.getBoundName(0);
                    if (boundName.equals(variables[0].getName())) {
                        kt = typeParams.get(0);
                        vt = _unknownType();
                        break;
                    } else if (boundName.equals(variables[1].getName())) {
                        kt = _unknownType();
                        vt = typeParams.get(0);
                        break;
                    }
                }
```

But I assume in scala and vavr, it is not _mapType that is the problem, but rather the *pattern*. In much of databind, we query the type along with the variable name (e.g. TypeBindings.findBoundType). But in some places like _mapType, we only query the type, and depend on the order to be correct and complete. I'm not sure this is even safe before this PR. But this is also exposed via JavaType.containedType which is used widely, so it's not viable to fix all use sites.

This PR takes an alternate approach. Instead of shrinking the TypeBindings for wildcards, it replaces the wildcard with `null`. I avoided this initially because not all use sites can handle null. So in this patch, I made the old TypeBindings methods return `unknownType` (i.e. `Object`) and made new ones that return `null`. This should be more compatible but still address the original bug. WDYT?