Allow default enums with @JsonCreator
This follows the pattern for READ_UNKNOWN_ENUM_VALUES_AS_NULL from here:
https://github.com/FasterXML/jackson-databind/pull/1642/files

but for READ_UNKNOWN_ENUM_VALUES_USING_DEFAULT_VALUE

It's more tricky because it wasn't immediately apparent how to get the
configured default value on the Creator impl side. I decided to try
using the EnumResolver, which works, but am not sure what the
repercussions of that are fully though.

It works the same way, so only IllegalArgumentExceptions will trigger
default behavior, so if you have some other custom creator exception
logic that will be unaffected.